@@ -8,40 +8,42 @@
 import CLIlib
 import Foundation
 
-protocol AppleDownloaderProtocol {
+protocol AppleDownloaderProtocol: Sendable {
     func list(force: Bool) async throws -> DownloadList
     func download(file: DownloadList.File) async throws -> URLSessionDownloadTaskProtocol?
 }
 
+@MainActor
 class AppleDownloader: HTTPClient, AppleDownloaderProtocol {
-
+    
     // control the progress of the download
     // not private for testability
     var downloadTask: URLSessionDownloadTaskProtocol?
-
+    
     func download(file: DownloadList.File) async throws -> URLSessionDownloadTaskProtocol? {
-
-        guard let downloadDelegate = env.urlSessionDownload.downloadDelegate() else {
-            fatalError("This method requires an injected download delegate")
-        }
-
+        
         guard !file.remotePath.isEmpty,
-            !file.filename.isEmpty,
-            file.fileSize > 0
+              !file.filename.isEmpty,
+              file.fileSize > 0
         else {
             log.error("🛑 Invalid file specification : \(file)")
             throw DownloadError.invalidFileSpec
         }
-
+        
         let fileURL = "https://developer.apple.com/services-account/download?path=\(file.remotePath)"
-
-        // pass a progress update client to the download delegate to receive progress updates
-        downloadDelegate.totalFileSize = file.fileSize
-        downloadDelegate.dstFilePath = URL(
-            fileURLWithPath: env.fileHandler.downloadFilePath(file: file)
-        )
-        downloadDelegate.startTime = Date.now
-
+        
+        let fh = self.env.fileHandler
+        //filehandler is not MainActor-isolated so we need to use Task { } to get the file path
+        let filePath = await Task {
+            await URL(
+                fileURLWithPath: fh.downloadFilePath(file: file)
+            )
+        }.value
+        let urlSessionDownload = self.env.urlSessionDownload(dstFilePath: filePath, totalFileSize: file.fileSize, startTime: Date.now)
+        guard let downloadDelegate = urlSessionDownload.downloadDelegate() else {
+            fatalError("This method requires an injected download delegate")
+        }
+        
         // make a call to start the download
         // first call, should send a redirect and an auth cookie
         self.downloadTask = try await downloadCall(url: fileURL, requestHeaders: ["Accept": "*/*"])
@@ -49,10 +51,10 @@ class AppleDownloader: HTTPClient, AppleDownloaderProtocol {
             dlt.resume()
             downloadDelegate.sema.wait()
         }
-
+        
         // returns when the download is completed
         return self.downloadTask
-
+        
     }
-
+    
 }
