@@ -7,17 +7,18 @@
 
 import CLIlib
 import Foundation
+import Subprocess
 
 // MARK: Command Line Tools
 // Command Line Tools installation functions
 extension ShellInstaller {
-
-    func installCommandLineTools(atPath file: URL) throws {
+    
+    func installCommandLineTools(atPath file: URL) async  throws {
 
         let filePath = file.path
 
         // check if file exists
-        guard env.fileHandler.fileExists(file: file, fileSize: 0) else {
+        guard self.env.fileHandler.fileExists(file: file, fileSize: 0) else {
             log.error("Command line disk image does not exist : \(filePath)")
             throw InstallerError.fileDoesNotExistOrIncorrect
         }
@@ -26,15 +27,15 @@ extension ShellInstaller {
         let totalSteps = 3
         var currentStep: Int = 0
 
-        var resultOptional: ShellOutput?
+        var result: ShellOutput
 
         // first mount the disk image
         log.debug("Mounting disk image \(file.lastPathComponent)")
         currentStep += 1
-        env.progressBar.update(step: currentStep, total: totalSteps, text: "Mounting disk image...")
-        resultOptional = try self.mountDMG(atURL: file)
-        if resultOptional == nil || resultOptional!.code != 0 {
-            log.error("Can not mount disk image : \(filePath)\n\(String(describing: resultOptional))")
+        self.env.progressBar.update(step: currentStep, total: totalSteps, text: "Mounting disk image...")
+        result = try await self.mountDMG(atURL: file)
+        if !result.terminationStatus.isSuccess {
+            log.error("Can not mount disk image : \(filePath)\n\(String(describing: result))")
             throw InstallerError.CLToolsInstallationError
         }
 
@@ -44,10 +45,10 @@ extension ShellInstaller {
         let pkgPath = pkg.path
         log.debug("Installing pkg \(pkgPath)")
         currentStep += 1
-        env.progressBar.update(step: currentStep, total: totalSteps, text: "Installing package...")
-        resultOptional = try self.installPkg(atURL: pkg)
-        if resultOptional == nil || resultOptional!.code != 0 {
-            log.error("Can not install package : \(pkgPath)\n\(String(describing: resultOptional))")
+        self.env.progressBar.update(step: currentStep, total: totalSteps, text: "Installing package...")
+        result = try await self.installPkg(atURL: pkg)
+        if !result.terminationStatus.isSuccess {
+            log.error("Can not install package : \(pkgPath)\n\(String(describing: result))")
             throw InstallerError.CLToolsInstallationError
         }
 
@@ -55,39 +56,33 @@ extension ShellInstaller {
         let mountedDiskImage = URL(fileURLWithPath: "/Volumes/Command Line Developer Tools")
         log.debug("Unmounting volume \(mountedDiskImage)")
         currentStep += 1
-        env.progressBar.update(step: currentStep, total: totalSteps, text: "Unmounting volume...")
-        resultOptional = try self.unmountDMG(volume: mountedDiskImage)
-        if resultOptional == nil || resultOptional!.code != 0 {
+        self.env.progressBar.update(step: currentStep, total: totalSteps, text: "Unmounting volume...")
+        result = try await self.unmountDMG(volume: mountedDiskImage)
+        if !result.terminationStatus.isSuccess {
             log.error(
-                "Can not unmount volume : \(mountedDiskImage)\n\(String(describing: resultOptional))"
+                "Can not unmount volume : \(mountedDiskImage)\n\(String(describing: result))"
             )
             throw InstallerError.CLToolsInstallationError
         }
     }
 
-    private func mountDMG(atURL dmg: URL) throws -> ShellOutput {
+    private func mountDMG(atURL dmg: URL) async throws -> ShellOutput {
 
         let dmgPath = dmg.path
 
         // check if file exists
-        guard env.fileHandler.fileExists(file: dmg, fileSize: 0) else {
+        guard self.env.fileHandler.fileExists(file: dmg, fileSize: 0) else {
             log.error("Disk Image does not exist : \(dmgPath)")
             throw InstallerError.fileDoesNotExistOrIncorrect
         }
 
         // hdiutil mount ./xcode-cli.dmg
-        let cmd = "\(HDIUTILCOMMAND) mount \"\(dmgPath)\""
-        let result = try env.shell.run(cmd)
-
-        return result
+        return try await run(.path(HDIUTILCOMMAND), arguments: ["mount", dmgPath])
     }
 
-    private func unmountDMG(volume: URL) throws -> ShellOutput {
+    private func unmountDMG(volume: URL) async throws -> ShellOutput {
 
         // hdiutil unmount /Volumes/Command\ Line\ Developer\ Tools/
-        let cmd = "\(HDIUTILCOMMAND) unmount \"\(volume.path)\""
-        let result = try env.shell.run(cmd)
-
-        return result
+        return try await run(.path(HDIUTILCOMMAND), arguments: ["unmount", volume.path])
     }
 }
