@@ -7,23 +7,22 @@
 
 import CLIlib
 import Foundation
+import Subprocess
 
 // MARK: PKG
 // generic PKG installation function
 extension ShellInstaller {
 
-    func installPkg(atURL pkg: URL) throws -> ShellOutput {
+    func installPkg(atURL pkg: URL) async throws -> ShellOutput {
 
         let pkgPath = pkg.path
 
         // check if file exists
-        guard env.fileHandler.fileExists(file: pkg, fileSize: 0) else {
+        guard self.env.fileHandler.fileExists(file: pkg, fileSize: 0) else {
             log.error("Package does not exist : \(pkgPath)")
             throw InstallerError.fileDoesNotExistOrIncorrect
         }
 
-        let cmd = "sudo \(INSTALLERCOMMAND) -pkg \"\(pkgPath)\" -target /"
-        let result = try env.shell.run(cmd)
-        return result
+        return try await run(.path(INSTALLERCOMMAND), arguments: ["-pkg", pkgPath, "-target", "/"])
     }
 }
