@@ -7,38 +7,39 @@
 
 import CLIlib
 import Foundation
+import Subprocess
 
 // MARK: XCODE
 // XCode installation functions
 extension ShellInstaller {
 
-    func installXcode(at src: URL) throws {
+    func installXcode(at src: URL) async throws {
 
         // unXIP, mv, 4 PKG to install
         let totalSteps = 2 + PKGTOINSTALL.count
         var currentStep: Int = 0
 
-        var resultOptional: CLIlib.ShellOutput?
+        var result: ShellOutput
 
         // first uncompress file
         log.debug("Decompressing files")
         // run synchronously as there is no output for this operation
         currentStep += 1
-        env.progressBar.update(
+        self.env.progressBar.update(
             step: currentStep,
             total: totalSteps,
             text: "Expanding Xcode xip (this might take a while)"
         )
-        resultOptional = try self.uncompressXIP(atURL: src)
-        if resultOptional == nil || resultOptional!.code != 0 {
-            log.error("Can not unXip file : \(resultOptional!)")
+        result = try await self.uncompressXIP(atURL: src)
+        if !result.terminationStatus.isSuccess {
+            log.error("Can not unXip file : \(result)")
             throw InstallerError.xCodeXIPInstallationError
         }
 
         // second move file to /Applications
         log.debug("Moving app to destination")
         currentStep += 1
-        env.progressBar.update(
+        self.env.progressBar.update(
             step: currentStep,
             total: totalSteps,
             text: "Moving Xcode to /Applications"
@@ -55,7 +56,7 @@ extension ShellInstaller {
         }
 
         let installedFile =
-            try self.moveApp(at: FileHandler.downloadDirectory.appendingPathComponent(appFile[0]))
+            try await self.moveApp(at: self.env.fileHandler.downloadDirectory().appendingPathComponent(appFile[0]))
 
         // /Applications/Xcode.app/Contents/Resources/Packages/
 
@@ -63,16 +64,16 @@ extension ShellInstaller {
         for pkg in PKGTOINSTALL {
             log.debug("Installing package \(pkg)")
             currentStep += 1
-            env.progressBar.update(
+            self.env.progressBar.update(
                 step: currentStep,
                 total: totalSteps,
                 text: "Installing additional packages... \(pkg)"
             )
-            resultOptional = try self.installPkg(
+            result = try await self.installPkg(
                 atURL: URL(fileURLWithPath: "\(installedFile)/Contents/resources/Packages/\(pkg)")
             )
-            if resultOptional == nil || resultOptional!.code != 0 {
-                log.error("Can not install pkg at : \(pkg)\n\(resultOptional!)")
+            if !result.terminationStatus.isSuccess {
+                log.error("Can not install pkg at : \(pkg)\n\(result)")
                 throw InstallerError.xCodePKGInstallationError
             }
         }
@@ -82,26 +83,23 @@ extension ShellInstaller {
     // expand a XIP file.  There is no way to create XIP file.
     // This code can not be tested without a valid, signed,  Xcode archive
     // https://en.wikipedia.org/wiki/.XIP
-    func uncompressXIP(atURL file: URL) throws -> ShellOutput {
+    func uncompressXIP(atURL file: URL) async throws -> ShellOutput {
 
         let filePath = file.path
 
         // not necessary, file existence has been checked before
-        guard env.fileHandler.fileExists(file: file, fileSize: 0) else {
+        guard self.env.fileHandler.fileExists(file: file, fileSize: 0) else {
             log.error("File to unXip does not exist : \(filePath)")
             throw InstallerError.fileDoesNotExistOrIncorrect
         }
 
         // synchronously uncompress in the download directory
-        let cmd =
-            "pushd \"\(FileHandler.downloadDirectory.path)\" && "
-            + "\(XIPCOMMAND) --expand \"\(filePath)\" && " + "popd"
-        let result = try env.shell.run(cmd)
+        let result = try await run(.name(XIPCOMMAND), arguments: ["--expand", filePath], workingDirectory: .init(env.fileHandler.downloadDirectory().path))
 
         return result
     }
 
-    func moveApp(at src: URL) throws -> String {
+    func moveApp(at src: URL) async throws -> String {
 
         // extract file name
         let fileName = src.lastPathComponent
@@ -111,7 +109,7 @@ extension ShellInstaller {
 
         log.debug("Going to move \n \(src) to \n \(appURL)")
         // move synchronously
-        try env.fileHandler.move(from: src, to: appURL)
+        try await self.env.fileHandler.move(from: src, to: appURL)
 
         return appURL.path
     }
