@@ -21,7 +21,7 @@ extension AppleDownloader {
 
         if !force {
             // load the list from file if we have it
-            downloadList = try? env.fileHandler.loadDownloadList()
+            downloadList = try? self.env.fileHandler.loadDownloadList()
         }
 
         if downloadList == nil {
@@ -50,7 +50,7 @@ extension AppleDownloader {
                 // grab authentication cookie for later download
                 if let cookies = response.value(forHTTPHeaderField: "Set-Cookie") {
                     // save the new cookies we received (ADCDownloadAuth)
-                    _ = try await env.secrets.saveCookies(cookies)
+                    _ = try await self.env.secrets.saveCookies(cookies)
                 } else {
                     // swiftlint:disable line_length
                     log.error(
@@ -61,7 +61,7 @@ extension AppleDownloader {
                 }
 
                 // success, save the list for reuse
-                _ = try env.fileHandler.saveDownloadList(list: downloadList!)
+                _ = try self.env.fileHandler.saveDownloadList(list: downloadList!)
 
             } else {
 
