@@ -7,13 +7,23 @@
 
 import CLIlib
 import Foundation
+import Logging
 
+@MainActor
 struct XCodeInstall {
-
+    
+    let log: Logger
+    let env: Environment
+    
+    public init(log: Logger = Log.defaultLogger(), env: Environment) {
+        self.log = log
+        self.env = env
+    }
+        
     // display a message to the user
     // avoid having to replicate the \n torough the code
     func display(_ msg: String, terminator: String = "\n") {
-        env.display.display(msg, terminator: terminator)
+        self.env.display.display(msg, terminator: terminator)
     }
 
 }
